+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/SecurableType.java
@@ -11,6 +11,9 @@ public enum SecurableType {
   @JsonProperty("catalog")
   CATALOG,
 
+  @JsonProperty("connection")
+  CONNECTION,
+
   @JsonProperty("external_location")
   EXTERNAL_LOCATION,