+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/SseEncryptionDetailsAlgorithm.java
@@ -1,0 +1,12 @@
+// Code generated from OpenAPI specs by Databricks SDK Generator. DO NOT EDIT.
+
+package com.databricks.sdk.service.catalog;
+
+import com.databricks.sdk.support.Generated;
+
+/** The type of key encryption to use (affects headers from s3 client). */
+@Generated
+public enum SseEncryptionDetailsAlgorithm {
+  AWS_SSE_KMS,
+  AWS_SSE_S3,
+}