+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/sharing/SharesAPI.java
@@ -6,7 +6,12 @@ import com.databricks.sdk.support.Generated;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-/** Databricks Shares REST API */
+/**
+ * A share is a container instantiated with :method:shares/create. Once created you can iteratively
+ * register a collection of existing data assets defined within the metastore using
+ * :method:shares/update. You can register data assets under their original name, qualified by their
+ * original schema, or provide alternate exposed names.
+ */
 @Generated
 public class SharesAPI {
   private static final Logger LOG = LoggerFactory.getLogger(SharesAPI.class);